backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==1.10.17
httpx==0.27.0
python-dotenv==1.0.1
websockets==12.0